✨ (slack_ai_agent): add conversation handler for Slack AI agent
Introduce a new module, `conversation.py`, to handle Slack
conversations. This refactor extracts conversation handling logic
from `event_handlers.py` to improve code organization and
maintainability. The new module includes functions to retrieve
thread history and process conversations, enhancing the Slack AI
agent's ability to respond to mentions and messages in threads.
